requirements/base.txt
Django==4.1.7
djangorestframework==3.14.0
django-environ==0.10.0
django-jazzmin==2.6.0
drf-yasg==1.21.5
django-ckeditor==6.5.1
Pillow==9.4.0
django-modeltranslation==0.18.9
django-phonenumber-field==7.0.2
psycopg2-binary==2.9.5
django-rosetta==0.9.8
redis==4.5.2
phonenumbers==8.13.7
<<<<<<< HEAD
django-filter==22.1
=======
sorl-thumbnail==12.9.0
django-filter==22.1
django-qrcode==0.3
pdf2image==1.16.3
weasyprint==58.1
qrcode==7.4.2
dj-rest-auth==3.0.0
django-allauth==0.52.0
sorl-thumbnail==12.9.0
django-filter==22.1
>>>>>>> e31a13b51cf0fd5c998e7e00946abe633c58926c
